+++ src/main/java/com/example/farmingholiday/controller/ListController.java
@@ -1,18 +1,15 @@
 package com.example.farmingholiday.controller;
 
-import com.example.farmingholiday.dto.BlockFarmingHolidayDto;
-import com.example.farmingholiday.dto.BlockHostDto;
-import com.example.farmingholiday.dto.BlockHouseDto;
-import com.example.farmingholiday.dto.FarmingHolidayDto;
+import com.example.farmingholiday.dto.home.BlockFarmingHolidayDto;
+import com.example.farmingholiday.dto.home.BlockHostDto;
+import com.example.farmingholiday.dto.home.BlockHouseDto;
 import com.example.farmingholiday.service.FarmingHolidayService;
 import com.example.farmingholiday.service.HostService;
 import com.example.farmingholiday.service.HouseService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @RequiredArgsConstructor
