+++ src/main/java/com/example/farmingholiday/dto/detail/DetailPageDto.java
@@ -1,4 +1,4 @@
-package com.example.farmingholiday.dto;
+package com.example.farmingholiday.dto.detail;
 
 import lombok.*;
