+++ src/main/java/com/example/farmingholiday/dto/detail/HouseDto.java
@@ -1,4 +1,4 @@
-package com.example.farmingholiday.dto;
+package com.example.farmingholiday.dto.detail;
 
 import com.example.farmingholiday.domain.House;
 import lombok.*;
