+++ src/main/java/com/example/farmingholiday/dto/home/HomePageDto.java
@@ -1,4 +1,4 @@
-package com.example.farmingholiday.dto;
+package com.example.farmingholiday.dto.home;
 
 import lombok.*;
