+++ src/main/java/com/example/farmingholiday/repository/LikeHostRepository.java
@@ -1,9 +1,7 @@
 package com.example.farmingholiday.repository;
 
 import com.example.farmingholiday.domain.Guest;
-import com.example.farmingholiday.domain.LikeFarmingHoliday;
 import com.example.farmingholiday.domain.LikeHost;
-import com.example.farmingholiday.dto.BlockHostDto;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
