+++ src/main/resources/application.yml
@@ -4,7 +4,7 @@ spring:
 
   jpa:
     hibernate:
-      ddl-auto: update
+      ddl-auto: create
     generate-ddl: true
     show-sql: false
     database: mysql
@@ -18,4 +18,4 @@ spring:
 
   sql:
     init:
-      mode: never
+      mode: always
